+++ cmd/xmtpd-cli/commands/rate_registry.go
@@ -203,15 +203,10 @@ func setupRateRegistryAdmin(
 		return nil, fmt.Errorf("could not create signer: %w", err)
 	}
 
-	paramAdmin, err := blockchain.NewParameterAdmin(logger, chainClient, signer, contracts)
-	if err != nil {
-		return nil, fmt.Errorf("could not create parameter admin: %w", err)
-	}
-
 	registryAdmin, err := blockchain.NewRatesAdmin(
 		logger,
-		paramAdmin,
 		chainClient,
+		signer,
 		contracts,
 	)
 	if err != nil {
