+++ pkg/config/blockchain.go
@@ -2,8 +2,9 @@ package config
 
 type ChainConfig struct {
 	AppChainDeploymentBlock          int    `json:"appChainDeploymentBlock"`
-	AppChainID                       int    `json:"appChainId"`
 	AppChainGateway                  string `json:"appChainGateway"`
+	AppChainID                       int    `json:"appChainId"`
+	AppChainParameterRegistry		 string `json:"appChainParameterRegistry"`
 	DistributionManager              string `json:"distributionManager"`
 	GroupMessageBroadcaster          string `json:"groupMessageBroadcaster"`
 	IdentityUpdateBroadcaster        string `json:"identityUpdateBroadcaster"`
