+++ pkg/payerreport/workers/integration_test.go
@@ -76,15 +76,11 @@ func setupBlockchain(
 	)
 	require.NoError(t, err)
 
-	parameterAdmin, err := blockchain.NewParameterAdmin(log, client, signer, contractsOptions)
-	require.NoError(t, err)
-
 	registryAdmin, err := blockchain.NewNodeRegistryAdmin(
 		log,
 		client,
 		signer,
 		contractsOptions,
-		parameterAdmin,
 	)
 	require.NoError(t, err)
