+++ pkg/testutils/config.go
@@ -45,6 +45,7 @@ func NewContractsOptions(t *testing.T, rpcURL, wsURL string) config.ContractsOpt
 			IdentityUpdateBroadcasterAddress: chainConfig.IdentityUpdateBroadcaster,
 			BackfillBlockPageSize:            500,
 			GatewayAddress:                   chainConfig.AppChainGateway,
+			ParameterRegistryAddress: chainConfig.AppChainParameterRegistry,
 		},
 		SettlementChain: config.SettlementChainOptions{
 			RPCURL:                      rpcURL,
